+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotImportAnnotator.java
@@ -15,7 +15,7 @@ public class RobotImportAnnotator implements Annotator {
 
     @Override
     public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {
-        if (!element.isValid() || !(element instanceof Import importElement)) {
+        if (!(element instanceof Import importElement)) {
             return;
         }