+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotKeywordAnnotator.java
@@ -1,8 +1,10 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.ide.inspections.compilation;
 
+import com.intellij.psi.util.PsiTreeUtil;
 import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTokenTypes;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordInvokable;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.VariableDefinitionGroup;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref.PyElementDeprecatedVisitor;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref.PyElementParentTraversalVisitor;
 import com.intellij.codeInspection.ProblemHighlightType;
@@ -19,14 +21,21 @@ public class RobotKeywordAnnotator implements Annotator {
 
     @Override
     public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {
-        if (!element.isValid() || !(element instanceof KeywordInvokable)) {
+        if (!(element instanceof KeywordInvokable)) {
             return;
         }
+
         PsiReference reference = element.getReference();
-        PsiElement resolvedElement = reference != null ? reference.resolve() : null;
+        PsiElement resolvedElement = reference.resolve();
         boolean isResolved = resolvedElement != null;
         if (!isResolved && element.getNode().getElementType() != RobotTokenTypes.SYNTAX_MARKER) {
-            holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.keyword.not-found")).range(element).create();
+            VariableDefinitionGroup variableDefinitionGroup = PsiTreeUtil.getParentOfType(element, VariableDefinitionGroup.class);
+            // Don't mark it as an unresolved keyword as it is an argument of a variable definition (marked by RobotVariableArgumentAnnotator).
+            // While lexing and parsing it isn't possible to differentiate between a keyword statement and an argument. Better say it as a keyword statement
+            // and change the highlighting and suppress every message than to try to modify the tree after lexing/parsing
+            if (variableDefinitionGroup == null) {
+                holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.keyword.not-found")).range(element).create();
+            }
         } else if (resolvedElement instanceof PyElement pyElement) {
             PyElementDeprecatedVisitor pyElementDeprecatedVisitor = new PyElementDeprecatedVisitor();
             PyElementVisitor pyElementParentTraversalVisitor = new PyElementParentTraversalVisitor(pyElementDeprecatedVisitor);