+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotVariableAnnotator.java
@@ -1,9 +1,5 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.ide.inspections.compilation;
 
-import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordStatement;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.PositionalArgument;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Variable;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.lang.annotation.HighlightSeverity;
@@ -12,6 +8,10 @@ import com.intellij.openapi.project.DumbAware;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiReference;
 import com.intellij.psi.util.PsiTreeUtil;
+import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordStatement;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.PositionalArgument;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Variable;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
@@ -20,11 +20,11 @@ import java.util.regex.Pattern;
 
 public class RobotVariableAnnotator implements Annotator, DumbAware {
 
-    private static final Pattern VARIABLE_PATTERN = Pattern.compile("^\\$\\{(.*)}$");
+    private static final Pattern VARIABLE_PATTERN = Pattern.compile("^\\$\\{(.*?)}$");
 
     @Override
     public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {
-        if (!element.isValid() || !(element instanceof Variable variable) || variable.isEmpty() || variable.isNested()) {
+        if (!(element instanceof Variable variable) || variable.isEmpty() || variable.isNested() || variable.isEnvironmentVariable()) {
             return;
         }
         PsiReference reference = element.getReference();
@@ -43,15 +43,13 @@ public class RobotVariableAnnotator implements Annotator, DumbAware {
             }
         }
 
-        if (element.isValid()) {
-            KeywordStatement keywordStatement = PsiTreeUtil.getParentOfType(element, KeywordStatement.class);
-            if (keywordStatement != null) {
-                List<PositionalArgument> positionalArguments = keywordStatement.getPositionalArguments();
-                if (keywordStatement.getGlobalVariable() != null && positionalArguments.size() > 1 && element == positionalArguments.getFirst()) {
-                    return;
-                }
+        KeywordStatement keywordStatement = PsiTreeUtil.getParentOfType(element, KeywordStatement.class);
+        if (keywordStatement != null) {
+            List<PositionalArgument> positionalArguments = keywordStatement.getPositionalArguments();
+            if (keywordStatement.getGlobalVariable() != null && positionalArguments.size() > 1 && element == positionalArguments.getFirst()) {
+                return;
             }
-            holder.newAnnotation(HighlightSeverity.WEAK_WARNING, RobotBundle.getMessage("annotation.variable.not-found")).range(element).create();
         }
+        holder.newAnnotation(HighlightSeverity.WEAK_WARNING, RobotBundle.getMessage("annotation.variable.not-found")).range(element).create();
     }
 }