+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/KeywordInvokable.java
@@ -1,6 +1,7 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.psi.element;
 
 import com.intellij.psi.PsiNamedElement;
+import com.intellij.psi.PsiReference;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
@@ -16,4 +17,8 @@ public interface KeywordInvokable extends RobotStatement, PsiNamedElement {
 
    @NotNull
    Collection<PositionalArgument> getPositionalArguments();
+
+   @NotNull
+   @Override
+   PsiReference getReference();
 }