+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/Variable.java
@@ -6,11 +6,9 @@ import com.intellij.psi.StubBasedPsiElement;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.stub.element.VariableStub;
 import org.jetbrains.annotations.NotNull;
 
-public interface Variable extends RobotStatement, PsiNamedElement, StubBasedPsiElement<VariableStub> {
+public interface Variable extends RobotStatement, PsiNamedElement, StubBasedPsiElement<VariableStub>, VariableName {
 
    boolean isNested();
-
-   boolean isEmpty();
 
    @NotNull
    String getName();
@@ -18,12 +16,4 @@ public interface Variable extends RobotStatement, PsiNamedElement, StubBasedPsiE
    @NotNull
    @Override
    PsiReference getReference();
-
-   default String getUnwrappedName() {
-      if (isEmpty()) {
-         return "";
-      }
-      String name = getName();
-      return name.substring(2, name.length() - 1);
-   }
 }