+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/VariableDefinitionGroup.java
@@ -4,5 +4,9 @@ import java.util.Collection;
 
 public interface VariableDefinitionGroup extends RobotStatement {
 
+    default boolean hasMoreThanOneVariable() {
+        return getDefinedVariables().size() > 1;
+    }
+
     Collection<DefinedVariable> getDefinedVariables();
 }